+++ src/compas_ui/rhino/install.py
@@ -8,6 +8,7 @@ import compas
 from compas.plugins import plugin
 import compas_rhino
 from compas_rhino.install import install as install_packages
+from compas_rhino.install import _filter_installable_packages
 from compas_rhino.install_plugin import install_plugin
 
 
@@ -20,35 +21,38 @@ def installable_rhino_packages():
 
 
 def install(version="7.0"):
-    plugin = os.path.join(HERE, "ui", "COMPAS")
+    plugin_name = "COMPAS"
+    plugin_path = os.path.join(HERE, "ui", plugin_name)
+    if not os.path.isdir(plugin_path):
+        raise Exception("Cannot find the plugin: {}".format(plugin_path))
 
-    if not os.path.isdir(plugin):
-        raise Exception("Cannot find the plugin: {}".format(plugin))
-
-    plugin_dir = os.path.abspath(plugin)
-
-    plugin_path, plugin_name = os.path.split(plugin_dir)
-    plugin_path = plugin_path or os.getcwd()
-    plugin_dev = os.path.join(plugin_dir, "dev")
+    plugin_path = os.path.abspath(plugin_path)
+    plugin_dev = os.path.join(plugin_path, "dev")
     plugin_config = os.path.join(plugin_dev, "config.json")
 
     with open(plugin_config, "r") as f:
         config = json.load(f)
 
-    if "packages" in config:
-        install_packages(version=version, packages=config["packages"])
+    packages = []
+    packages = _filter_installable_packages(version, packages)
 
-    install_plugin(plugin)
+    if "packages" in config:
+        for name in config["packages"]:
+            if name not in packages:
+                packages.append(name)
+
+    install_packages(version=version, packages=packages)
+    install_plugin(plugin_path)
 
     if compas.WINDOWS:
-        ruipy = os.path.join(plugin_dev, "rui.py")
-        ruiname = "{}.rui".format(plugin_name)
+        plugin_ruipy = os.path.join(plugin_dev, "rui.py")
+        plugin_rui = "{}.rui".format(plugin_name)
         python_plugins_path = compas_rhino._get_rhino_pythonplugins_path(version)
 
-        call(sys.executable + " " + ruipy, shell=True)
+        call(sys.executable + " " + plugin_ruipy, shell=True)
         copyfile(
-            os.path.join(plugin_dev, ruiname),
-            os.path.join(python_plugins_path, "..", "..", "UI", ruiname),
+            os.path.join(plugin_dev, plugin_rui),
+            os.path.join(python_plugins_path, "..", "..", "UI", plugin_rui),
         )
 
 
@@ -59,7 +63,10 @@ if __name__ == "__main__":
         description="COMPAS UI installer command line utility."
     )
     parser.add_argument(
-        "--version", default="7.0", choices=["6.0", "7.0"], help="Version of Rhino."
+        "--version",
+        default="7.0",
+        choices=["6.0", "7.0"],
+        help="Version of Rhino.",
     )
     args = parser.parse_args()