+++ src/compas_ui/rhino/ui/COMPAS/dev/COMPAS_init_cmd.py
@@ -1,11 +1,8 @@
 import os
 import json
-from compas_ui.ui import UI
 
-# TODO: auto start the cloud
-# TODO: change splash screen
-# TODO: check why loading takes so long
-# TODO: check if current environment has compas_ui
+from compas_ui.ui import UI
+from compas_ui.rhino.forms import SplashForm
 
 
 __commandname__ = "COMPAS_init"
@@ -23,8 +20,10 @@ def RunCommand(is_interactive):
 
     UI.reset()
     ui = UI(config=CONFIG)
-    ui.splash(url=SPLASH)
     ui.scene_clear()
+
+    browser = SplashForm(title=ui.name, url=SPLASH)
+    browser.show()
 
 
 if __name__ == "__main__":