+++ src/compas_ui/rhino/ui/COMPAS/dev/COMPAS_mesh_create_cmd.py
@@ -1,7 +1,5 @@
 from compas_ui.ui import UI
 
-# TODO: disable recording for now
-
 
 __commandname__ = "COMPAS_mesh_create"