+++ src/compas_ui/rhino/ui/COMPAS/dev/COMPAS_scene_objects_cmd.py
@@ -1,10 +1,5 @@
 from compas_ui.ui import UI
 
-# TODO: add namespace per plugin
-# TODO: root objects under namespaces
-# TODO: provide COMPAS as default namespace
-# TODO: links instead of buttons?
-
 
 __commandname__ = "COMPAS_scene_objects"
 
@@ -14,7 +9,6 @@ def RunCommand(is_interactive):
 
     ui = UI()
     ui.scene_objects()
-    ui.record()
 
 
 if __name__ == "__main__":