+++ src/compas_ui/rhino/ui/COMPAS/dev/__plugin__.py
@@ -1,3 +1,6 @@
+# don't modify the contents of this file
+# it is auto-generated....
+
 id = "{6c18fcab-3ae5-4b03-b515-d33d8078b977}"
 version = "0.0.1.0"
 title = "COMPAS"