+++ src/compas_ui/rhino/ui/COMPAS/dev/rui.py
@@ -2,9 +2,9 @@ import os
 from compas_ui.rhino.rui import Rui
 
 HERE = os.path.dirname(__file__)
-CONFIGPATH = os.path.join(HERE, "config.json")
-RUIPATH = os.path.join(HERE, "COMPAS.rui")
+UI = os.path.join(HERE, "ui.json")
+RUI = os.path.join(HERE, "COMPAS.rui")
 
-rui = Rui.from_json(CONFIGPATH, RUIPATH)
+rui = Rui.from_json(UI, RUI)
 
 rui.write()