+++ CHANGELOG.md
@@ -18,11 +18,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Added `compas_ui.rhino.forms.SearchPathsForm`.
 * Added default controller for all non-system commands.
 * Added `active_object` to `scene`.
+* Added `use_tab` option to `compas_ui.rhino.forms.SettingsForm`.
+* Added `export` and `remove` buttons to `compas_ui.rhino.forms.SceneForm`.
+* Added `compas_ui.rhino.forms.ToolbarForm`.
 * Added `frame` to `compas_ui.objects.Object`.
 
 ### Changed
 
 * Change on `SceneObjectsForm` is reflected realtime.
+* Locked `sphinx` to `4.5`
 
 ### Removed